part1_basics_langgraph/3_prompt_template.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model='gpt-4.1-mini', temperature=0)

from langchain_core.prompts import PromptTemplate

# Prompt template
template = PromptTemplate(
    template="""

    Please summarize the research paper titled "{paper_input}" with the following specifications:
    Explanation Style: {style_input}  
    Explanation Length: {length_input}  
    1. Mathematical Details:  
    - Include relevant mathematical equations if present in the paper.  
    - Explain the mathematical concepts using simple, intuitive code snippets where applicable.  
    2. Analogies:  
    - Use relatable analogies to simplify complex ideas.  
    If certain information is not available in the paper, respond with: "Insufficient information available" instead of guessing.  
    Ensure the summary is clear, accurate, and aligned with the provided style and length.
    """,
    input_variables=['paper_input', 'style_input','length_input'],
    validate_template=True
)

chain = template | model

result = chain.invoke({
    'paper_input':'Attention is all you need',
    'style_input':'Simple and intuitive',
    'length_input':'500 words'
})

print(result.content)